// src/shared/enums/errors.enum.ts
export enum ErrorsEnum {
    NEWS_NOT_FOUND = 'NEWS_NOT_FOUND',
    USER_NOT_FOUND = 'USER_NOT_FOUND',
    NOT_A_STRONG_PASSWORD = 'NOT_A_STRONG_PASSWORD',
    INCORRECT_LOGIN = 'INCORRECT_LOGIN',
    INCORRECT_EMAIL = 'INCORRECT_EMAIL',
    USER_WITH_THIS_EMAIL_ALREADY_EXIST = 'USER_WITH_THIS_EMAIL_ALREADY_EXIST',
    USER_WITH_THIS_LOGIN_ALREADY_EXIST = 'USER_WITH_THIS_LOGIN_ALREADY_EXIST',
    REFRESH_TOKEN_MALFORMED = 'REFRESH_TOKEN_MALFORMED',
    ACCESS_DENIED = 'ACCESS_DENIED',
    INCORRECT_EMAIL_OR_PASSWORD = 'INCORRECT_EMAIL_OR_PASSWORD',
}